+++ OpenAVRc/pulses/pulses_avr.h
@@ -40,7 +40,7 @@ extern uint8_t moduleFlag[NUM_MODULES];
 #define MAX_MIXER_DELTA (50*16) /* 50ms max as an interval between 2 mixer calculations */
 #define SETUP_PULSES_DURATION 1000 // 500us
 
-extern uint8_t dt;
+extern uint16_t dt;
 
 void startPulses();
 inline bool pulsesStarted()
